Rename cypress to system
