+++ .gitignore
@@ -99,10 +99,10 @@ RoboFile.ini
 /media/com_media/css/mediamanager.min.css.map
 
 #cypress
-/tests/cypress/output/screenshots
-!/tests/cypress/output/screenshots/.gitkeep
-/tests/cypress/output/videos
-!/tests/cypress/output/videos/.gitkeep
+/tests/System/output/screenshots
+!/tests/System/output/screenshots/.gitkeep
+/tests/System/output/videos
+!/tests/System/output/videos/.gitkeep
 cypress.config.js
 
 # WebAuthn FIDO metadata cache
