--- configuration.php.bak
@@ -1,109 +1,0 @@
-<?php
-class JConfig {
-	public $offline = '0';
-	public $offline_message = 'This site is down for maintenance.<br />Please check back again soon.';
-	public $display_offline_message = '1';
-	public $offline_image = '';
-	public $sitename = 'Cypress';
-	public $editor = 'tinymce';
-	public $captcha = '0';
-	public $list_limit = '20';
-	public $access = '1';
-	public $debug = '1';
-	public $debug_lang = '0';
-	public $debug_lang_const = '1';
-	public $dbtype = 'pgsql';
-	public $host = 'postgres-test';
-	public $user = 'root';
-	public $password = 'root';
-	public $db = 'joomla_cms4';
-	public $dbprefix = 'j_';
-	public $dbencryption = '0';
-	public $dbsslverifyservercert = '';
-	public $dbsslkey = '';
-	public $dbsslcert = '';
-	public $dbsslca = '';
-	public $dbsslcipher = '';
-	public $live_site = '';
- public $secret = 'tEstValue';
-	public $gzip = '0';
-	public $error_reporting = 'development';
-	public $helpurl = 'https://help.joomla.org/proxy?keyref=Help{major}{minor}:{keyref}&lang={langcode}';
-	public $ftp_host = '';
-	public $ftp_port = '';
-	public $ftp_user = '';
-	public $ftp_pass = '';
-	public $ftp_root = '';
-	public $ftp_enable = '0';
-	public $offset = 'UTC';
-	public $mailonline = '1';
-	public $mailer = 'smtp';
-	public $mailfrom = 'admin@example.com';
-	public $fromname = 'Cypress';
-	public $sendmail = '/usr/sbin/sendmail';
-	public $smtpauth = '0';
-	public $smtpuser = '';
-	public $smtppass = '';
-	public $smtphost = 'mailcatcher-test';
-	public $smtpsecure = 'none';
-	public $smtpport = '1025';
-	public $caching = '0';
-	public $cache_handler = 'file';
-	public $cachetime = '15';
-	public $cache_platformprefix = '0';
-	public $MetaDesc = '';
-	public $MetaTitle = '1';
-	public $MetaAuthor = '1';
-	public $MetaVersion = '0';
-	public $robots = '';
-	public $sef = '1';
-	public $sef_rewrite = '1';
-	public $sef_suffix = '0';
-	public $unicodeslugs = '0';
-	public $feed_limit = '10';
-	public $feed_email = 'none';
-	public $log_path = '/var/www/html/administrator/logs';
-	public $tmp_path = '/var/www/html/tmp';
-	public $lifetime = '9999';
-	public $session_handler = 'database';
-	public $shared_session = '0';
-	public $session_metadata = '1';
-	public $memcache_persist = '1';
-	public $memcache_compress = '0';
-	public $memcache_server_host = 'localhost';
-	public $memcache_server_port = '11211';
-	public $memcached_persist = '1';
-	public $memcached_compress = '0';
-	public $memcached_server_host = 'localhost';
-	public $memcached_server_port = '11211';
-	public $redis_persist = '1';
-	public $redis_server_host = 'localhost';
-	public $redis_server_port = '6379';
-	public $redis_server_auth = '';
-	public $redis_server_db = '0';
-	public $proxy_enable = '0';
-	public $proxy_host = '';
-	public $proxy_port = '';
-	public $proxy_user = '';
-	public $proxy_pass = '';
-	public $massmailoff = '0';
-	public $replyto = '';
-	public $replytoname = '';
-	public $MetaKeys = '';
-	public $MetaRights = '';
-	public $sitename_pagetitles = '0';
-	public $force_ssl = '0';
-	public $session_memcache_server_host = 'localhost';
-	public $session_memcache_server_port = '11211';
-	public $session_memcached_server_host = 'localhost';
-	public $session_memcached_server_port = '11211';
-	public $session_redis_persist = '1';
-	public $session_redis_server_host = 'localhost';
-	public $session_redis_server_port = '6379';
-	public $session_redis_server_auth = '';
-	public $session_redis_server_db = '0';
-	public $frontediting = '1';
-	public $cookie_domain = '';
-	public $cookie_path = '';
-	public $asset_id = '1';
-}
