+++ cypress.config.dist.js
@@ -1,10 +1,10 @@
 const { defineConfig } = require('cypress');
-const setupPlugins = require('./tests/cypress/plugins/index');
+const setupPlugins = require('./tests/System/plugins/index');
 
 module.exports = defineConfig({
-  fixturesFolder: 'tests/cypress/fixtures',
-  videosFolder: 'tests/cypress/output/videos',
-  screenshotsFolder: 'tests/cypress/output/screenshots',
+  fixturesFolder: 'tests/System/fixtures',
+  videosFolder: 'tests/System/output/videos',
+  screenshotsFolder: 'tests/System/output/screenshots',
   viewportHeight: 1000,
   viewportWidth: 1200,
   e2e: {
@@ -13,12 +13,12 @@ module.exports = defineConfig({
     },
     baseUrl: 'http://localhost/',
     specPattern: [
-      'tests/cypress/integration/install/**/*.cy.{js,jsx,ts,tsx}',
-      'tests/cypress/integration/administrator/**/*.cy.{js,jsx,ts,tsx}',
-      'tests/cypress/integration/site/**/*.cy.{js,jsx,ts,tsx}',
-      'tests/cypress/integration/api/**/*.cy.{js,jsx,ts,tsx}',
+      'tests/System/integration/install/**/*.cy.{js,jsx,ts,tsx}',
+      'tests/System/integration/administrator/**/*.cy.{js,jsx,ts,tsx}',
+      'tests/System/integration/site/**/*.cy.{js,jsx,ts,tsx}',
+      'tests/System/integration/api/**/*.cy.{js,jsx,ts,tsx}',
     ],
-    supportFile: 'tests/cypress/support/index.js',
+    supportFile: 'tests/System/support/index.js',
     scrollBehavior: 'center',
     browser: 'firefox',
     screenshotOnRunFailure: true,
