+++ tests/README.md
@@ -17,4 +17,4 @@ The javascript tests test the Joomla-specific Javascript code. For further infor
 
 Cypress tests
 ==========
-The Cypress tests test the user interface in a real browser and the webservices API of Joomla with the help of cypress.io. For further information on the tests and on how to run them, please check out [tests/cypress/README.md](cypress/README.md).
+The Cypress tests test the user interface in a real browser and the webservices API of Joomla with the help of cypress.io. For further information on the tests and on how to run them, please check out [tests/System/README.md](cypress/README.md).
