+++ tests/System/data/com_media/test-image-1.jpg
