+++ tests/System/data/com_media/test-image-2.png
