+++ tests/System/integration/administrator/components/com_content/ContentCategoryList.cy.js
