+++ tests/System/integration/api/com_banners/Banners.cy.js
