+++ tests/System/integration/api/com_media/Adapters.cy.js
