+++ tests/System/integration/api/com_media/Files.cy.js
@@ -62,7 +62,7 @@ describe('Test that media files API endpoint', () => {
   });
 
   it('can create a file without adapter', () => {
-    cy.readFile('tests/cypress/data/com_media/test-image-1.jpg', 'binary')
+    cy.readFile('tests/System/data/com_media/test-image-1.jpg', 'binary')
       .then((data) => cy.api_post('/media/files', { path: 'test-dir/test.jpg', content: Buffer.from(data, 'binary').toString('base64') }))
       .then((response) => {
         cy.wrap(response).its('body').its('data').its('attributes')
@@ -87,7 +87,7 @@ describe('Test that media files API endpoint', () => {
   });
 
   it('can create a file with adapter', () => {
-    cy.readFile('tests/cypress/data/com_media/test-image-1.jpg', 'binary')
+    cy.readFile('tests/System/data/com_media/test-image-1.jpg', 'binary')
       .then((data) => cy.api_post('/media/files', { path: 'local-images:/test-dir/test.jpg', content: Buffer.from(data, 'binary').toString('base64') }))
       .then((response) => {
         cy.wrap(response).its('body').its('data').its('attributes')
@@ -113,7 +113,7 @@ describe('Test that media files API endpoint', () => {
 
   it('can update a file without adapter', () => {
     cy.task('writeFile', { path: 'images/test-dir/override.jpg', content: '1' })
-      .then(() => cy.readFile('tests/cypress/data/com_media/test-image-1.jpg', 'binary'))
+      .then(() => cy.readFile('tests/System/data/com_media/test-image-1.jpg', 'binary'))
       .then((data) => cy.api_patch(
         '/media/files/test-dir/override.jpg',
         { path: 'test-dir/override.jpg', content: Buffer.from(data, 'binary').toString('base64') },
@@ -142,7 +142,7 @@ describe('Test that media files API endpoint', () => {
 
   it('can update a file with adapter', () => {
     cy.task('writeFile', { path: 'images/test-dir/override.jpg', content: '1' })
-      .then(() => cy.readFile('tests/cypress/data/com_media/test-image-1.jpg', 'binary'))
+      .then(() => cy.readFile('tests/System/data/com_media/test-image-1.jpg', 'binary'))
       .then((data) => cy.api_patch(
         '/media/files/local-images:/test-dir/override.jpg',
         { path: 'local-images:/test-dir/override.jpg', content: Buffer.from(data, 'binary').toString('base64') },
