+++ tests/System/integration/site/components/com_content/Category.cy.js
