+++ tests/System/integration/site/components/com_content/Featured.cy.js
